web/requirements.txt
Django
requests
psycopg2
mysqlclient
djangorestframework